requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
Deprecated==1.2.13
google-auth==2.1.0
idna==3.2
kubernetes==18.20.0
mysql-connector-python==8.0.26
oauthlib==3.1.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
websocket-client==1.2.1
wrapt==1.12.1
